[*] ID使用雪花算法

+++ ingot-framework/ingot-core/src/main/java/com/ingot/framework/core/constants/TenantConstants.java
@@ -16,5 +16,5 @@ public interface TenantConstants {
     /**
      * 默认租户ID
      */
-    Integer DEFAULT_TENANT_ID = 1;
+    Long DEFAULT_TENANT_ID = 1L;
 }
