+++ ingot-framework/ingot-security/src/main/java/com/ingot/framework/security/oauth2/jwt/JwtClaimNamesExtension.java
@@ -25,16 +25,16 @@ public interface JwtClaimNamesExtension extends JwtClaimNames {
         return MapUtil.get(source.getClaims(), SUB, String.class);
     }
 
-    static Integer getId(Jwt source) {
-        return MapUtil.get(source.getClaims(), JwtClaimNamesExtension.ID, Integer.class);
+    static Long getId(Jwt source) {
+        return MapUtil.get(source.getClaims(), JwtClaimNamesExtension.ID, Long.class);
     }
 
-    static Integer getDept(Jwt source) {
-        return MapUtil.get(source.getClaims(), JwtClaimNamesExtension.DEPT, Integer.class);
+    static Long getDept(Jwt source) {
+        return MapUtil.get(source.getClaims(), JwtClaimNamesExtension.DEPT, Long.class);
     }
 
-    static Integer getTenantId(Jwt source) {
-        return MapUtil.get(source.getClaims(), JwtClaimNamesExtension.TENANT, Integer.class);
+    static Long getTenantId(Jwt source) {
+        return MapUtil.get(source.getClaims(), JwtClaimNamesExtension.TENANT, Long.class);
     }
 
     static String getAuthType(Jwt source) {
