+++ ingot-framework/ingot-tenant/src/main/java/com/ingot/framework/tenant/TenantEnv.java
@@ -25,8 +25,8 @@ public class TenantEnv {
      * @param tenantId 目标租户
      * @param func 执行操作
      */
-    public static void runAs(Integer tenantId, Run func) {
-        final Integer pre = TenantContextHolder.get();
+    public static void runAs(Long tenantId, Run func) {
+        final Long pre = TenantContextHolder.get();
         try {
             log.trace("TenantEnv - 切换租户环境 {} -> {}, 开始执行", pre, tenantId);
             TenantContextHolder.set(tenantId);
@@ -47,8 +47,8 @@ public class TenantEnv {
      * @param func 执行操作
      * @return 返回结果
      */
-    public static <T> T applyAs(Integer tenantId, Apply<T> func) {
-        final Integer pre = TenantContextHolder.get();
+    public static <T> T applyAs(Long tenantId, Apply<T> func) {
+        final Long pre = TenantContextHolder.get();
         try {
             log.trace("TenantEnv - 切换租户环境 {} -> {}, 开始执行", pre, tenantId);
             TenantContextHolder.set(tenantId);
