+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleAuthority.java
@@ -23,11 +23,11 @@ public class SysRoleAuthority extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 权限ID
      */
-    private Integer authorityId;
+    private Long authorityId;
 
 }
