+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/domain/SysRoleMenu.java
@@ -23,11 +23,11 @@ public class SysRoleMenu extends BaseModel<SysRoleAuthority> {
     /**
      * 角色ID
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 菜单ID
      */
-    private Integer menuId;
+    private Long menuId;
 
 }
