+++ ingot-service/ingot-pms/ingot-pms-api/src/main/java/com/ingot/cloud/pms/api/model/vo/user/UserProfileVO.java
@@ -18,12 +18,12 @@ public class UserProfileVO implements Serializable {
     /**
      * 部门ID
      */
-    private Integer deptId;
+    private Long deptId;
 
     /**
      * 角色ID
      */
-    private List<Integer> roleIds;
+    private List<Long> roleIds;
 
     /**
      * 用户名
