+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/Oauth2RegisteredClientService.java
@@ -26,7 +26,7 @@ public interface Oauth2RegisteredClientService extends BaseService<Oauth2Registe
      * @param roleIds 角色ID
      * @return Oauth2RegisteredClient 列表
      */
-    List<Oauth2RegisteredClient> getClientsByRoles(@Param("list") List<Integer> roleIds);
+    List<Oauth2RegisteredClient> getClientsByRoles(@Param("list") List<Long> roleIds);
 
     /**
      * 条件查询
