+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/SysMenuService.java
@@ -41,5 +41,5 @@ public interface SysMenuService extends BaseService<SysMenu> {
      *
      * @param id id
      */
-    void removeMenuById(int id);
+    void removeMenuById(long id);
 }
