+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/SysRoleOauthClientService.java
@@ -22,14 +22,14 @@ public interface SysRoleOauthClientService extends BaseService<SysRoleOauthClien
      *
      * @param params 关联参数
      */
-    void clientBindRoles(RelationDTO<Integer, Integer> params);
+    void clientBindRoles(RelationDTO<Long, Long> params);
 
     /**
      * 角色绑定客户端
      *
      * @param params 关联参数
      */
-    void roleBindClients(RelationDTO<Integer, Integer> params);
+    void roleBindClients(RelationDTO<Long, Long> params);
 
     /**
      * 获取角色客户端
@@ -40,7 +40,7 @@ public interface SysRoleOauthClientService extends BaseService<SysRoleOauthClien
      * @param condition 条件参数
      * @return 分页用户
      */
-    IPage<Oauth2RegisteredClient> getRoleClients(int roleId,
+    IPage<Oauth2RegisteredClient> getRoleClients(long roleId,
                                                  Page<?> page,
                                                  boolean isBind,
                                                  Oauth2RegisteredClient condition);
