+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/SysRoleUserService.java
@@ -25,7 +25,7 @@ public interface SysRoleUserService extends BaseService<SysRoleUser> {
      * @param userId 用户ID
      * @return 操作是否成功
      */
-    boolean removeByUserId(int userId);
+    boolean removeByUserId(long userId);
 
     /**
      * 更新用户角色，将用户角色更新为指定角色
@@ -33,21 +33,21 @@ public interface SysRoleUserService extends BaseService<SysRoleUser> {
      * @param userId 用户ID
      * @param roles  待设置的角色
      */
-    void updateUserRole(int userId, List<Integer> roles);
+    void updateUserRole(long userId, List<Long> roles);
 
     /**
      * 用户绑定角色
      *
      * @param params 关联参数
      */
-    void userBindRoles(RelationDTO<Integer, Integer> params);
+    void userBindRoles(RelationDTO<Long, Long> params);
 
     /**
      * 角色绑定用户
      *
      * @param params 关联参数
      */
-    void roleBindUsers(RelationDTO<Integer, Integer> params);
+    void roleBindUsers(RelationDTO<Long, Long> params);
 
     /**
      * 获取角色用户
@@ -58,7 +58,7 @@ public interface SysRoleUserService extends BaseService<SysRoleUser> {
      * @param condition 查询条件
      * @return 分页用户
      */
-    IPage<SysUser> getRoleUsers(int roleId,
+    IPage<SysUser> getRoleUsers(long roleId,
                                 Page<?> page,
                                 boolean isBind,
                                 SysUser condition);
