+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/impl/Oauth2RegisteredClientServiceImpl.java
@@ -50,7 +50,7 @@ public class Oauth2RegisteredClientServiceImpl extends BaseServiceImpl<Oauth2Reg
     private final SysRoleOauthClientService sysRoleOauthClientService;
 
     @Override
-    public List<Oauth2RegisteredClient> getClientsByRoles(List<Integer> roleIds) {
+    public List<Oauth2RegisteredClient> getClientsByRoles(List<Long> roleIds) {
         return getBaseMapper().getClientsByRoles(roleIds);
     }
